feat: update href in docs and try to fix not-found page
