+++ apps/ui-layout/app/robots.ts
@@ -5,94 +5,9 @@ export default function robots(): MetadataRoute.Robots {
     rules: [
       {
         userAgent: '*',
-        allow: [
-          '/',
-          '/components',
-          '/blocks',
-          '/api/og',
-          '/og',
-          '/_next/static',
-          '/favicon.ico',
-          '/robots.txt',
-          '/llms.txt',
-          '/llms-full.txt',
-          '/ai.txt',
-          '/sitemap.xml',
-        ],
-        disallow: [
-          '/admin',
-          '/api/admin',
-          '/_next',
-          '/static',
-          '/temp',
-          '/cache',
-          '/logs',
-          '/.env',
-          '/.git',
-          '/node_modules',
-          '/src',
-          '/public/temp',
-          '/*.json$',
-          '/search',
-          '/preview',
-          '/draft',
-        ],
-        crawlDelay: 1,
-      },
-      {
-        userAgent: 'Googlebot',
-        allow: ['/', '/components', '/blocks', '/api/og', '/og', '/llms.txt', '/llms-full.txt', '/ai.txt'],
-        disallow: ['/admin', '/api/admin', '/search', '/preview'],
-        crawlDelay: 0.5,
-      },
-      {
-        userAgent: 'Bingbot',
-        allow: ['/', '/components', '/blocks'],
-        disallow: ['/admin', '/api', '/search'],
-        crawlDelay: 1,
-      },
-      {
-        userAgent: 'GPTBot',
-        allow: ['/', '/components', '/blocks'],
-        disallow: ['/admin', '/api', '/search', '/preview'],
-        crawlDelay: 2,
-      },
-      {
-        userAgent: 'ChatGPT-User',
-        allow: '/',
-        disallow: ['/admin', '/api'],
-      },
-      {
-        userAgent: 'CCBot',
-        allow: ['/', '/components', '/blocks'],
-        disallow: ['/admin', '/api', '/search'],
-        crawlDelay: 3,
-      },
-      {
-        userAgent: 'anthropic-ai',
-        allow: ['/', '/components', '/blocks'],
-        disallow: ['/admin', '/api', '/search'],
-        crawlDelay: 2,
-      },
-      {
-        userAgent: 'Google-Extended',
-        allow: ['/', '/components', '/blocks'],
-        disallow: ['/admin', '/api/admin', '/search'],
-      },
-      {
-        userAgent: 'GoogleOther',
-        allow: ['/', '/components', '/blocks', '/og'],
-        disallow: ['/admin', '/api/admin', '/search'],
-      },
-      {
-        userAgent: 'Googlebot-Image',
-        allow: ['/og', '/components', '/blocks', '/public'],
-        disallow: ['/admin', '/temp'],
-      },
-      {
-        userAgent: 'Googlebot-Video',
         allow: ['/'],
-        disallow: ['/admin', '/api'],
+        disallow: [],
+        crawlDelay: 1,
       },
     ],
     sitemap: 'https://www.ui-layouts.com/sitemap.xml',
